[pulsetracker-0001] Move types to a global.d.ts file
