+++ packages/dashboard/pages/index.tsx
@@ -1,4 +1,5 @@
 import { utils } from '@pulse-tracker/utils';
+import { plugin } from '@pulse-tracker/plugin';
 import styled from '@emotion/styled';
 
 const StyledPage = styled.div`
@@ -7,6 +8,10 @@ const StyledPage = styled.div`
 `;
 
 export function Index() {
+  if (typeof window !== 'undefined') {
+    const rum = plugin.config({ id: "1", projectId: "miweb", urlEndpoint: "https://localhost/saveMetric" });
+    rum.measureFid((fid) => console.log("Mostramos el fid ", fid));
+  }
   const title = utils();
   return (
     <StyledPage>
