+++ packages/plugin/src/lib/global.d.ts
@@ -4,3 +4,9 @@ import { Metric } from 'web-vitals';
 export interface ProjectMetric extends Metric {
   projectId: string;
 }
+
+export interface PluginConfig {
+  id: string;
+  projectId: string;
+  urlEndpoint: string;
+}
